Update FTEs to accept deciaml values

+++ app/assets/javascripts/forms_ctrl.js
@@ -273,14 +273,14 @@
       $scope.fte_2 = 0;
       $scope.fte_3 = 0;
       for(var i = 0; i < $scope.challenge_formData.length; i++){
-            if ($scope.challenge_formData[i].fte_1 === parseInt($scope.challenge_formData[i].fte_1)) {
-              $scope.fte_1 += parseInt($scope.challenge_formData[i].fte_1);
+            if ($scope.challenge_formData[i].fte_1 === parseFloat($scope.challenge_formData[i].fte_1)) {
+              $scope.fte_1 += parseFloat($scope.challenge_formData[i].fte_1);
             }
-            if ($scope.challenge_formData[i].fte_2 === parseInt($scope.challenge_formData[i].fte_2)) {
-              $scope.fte_2 += parseInt($scope.challenge_formData[i].fte_2);
+            if ($scope.challenge_formData[i].fte_2 === parseFloat($scope.challenge_formData[i].fte_2)) {
+              $scope.fte_2 += parseFloat($scope.challenge_formData[i].fte_2);
             }
-            if ($scope.challenge_formData[i].fte_3 === parseInt($scope.challenge_formData[i].fte_3)) {
-              $scope.fte_3 += parseInt($scope.challenge_formData[i].fte_3);
+            if ($scope.challenge_formData[i].fte_3 === parseFloat($scope.challenge_formData[i].fte_3)) {
+              $scope.fte_3 += parseFloat($scope.challenge_formData[i].fte_3);
             }
           };
     };
